backend/requirements.txt
aiohappyeyeballs>=2.6,<2.7
aiohttp>=3.13,<3.14
aiolimiter>=1.2,<1.3
aiosignal>=1.4,<1.5
annotated-types>=0.7,<0.8
anyio>=4.12,<4.13
attrs>=25.4,<26.0
brightdata-sdk>=2.0,<2.1
certifi>=2026.2,<2027.0
charset-normalizer>=3.4,<3.5
click>=8.3,<8.4
distro>=1.9,<2.0
fastapi>=0.109,<0.110
filelock>=3.24,<3.25
frozenlist>=1.8,<1.9
h11>=0.16,<0.17
httpcore>=1.0,<1.1
httptools>=0.7,<0.8
httpx>=0.28,<0.29
idna>=3.11,<4.0
jiter>=0.13,<0.14
jsonpatch>=1.33,<1.34
jsonpointer>=3.0,<3.1
langchain-core>=1.2,<1.3
langchain-openai>=1.1,<1.2
langgraph>=1.0,<1.1
langgraph-checkpoint>=4.0,<4.1
langgraph-prebuilt>=1.0,<1.1
langgraph-sdk>=0.3,<0.4
langsmith>=0.7,<0.8
multidict>=6.7,<6.8
openai>=2.24,<2.25
orjson>=3.11,<3.12
ormsgpack>=1.12,<1.13
packaging>=26.0,<27.0
propcache>=0.4,<0.5
pydantic>=2.12,<2.13
pydantic-settings>=2.2,<2.3
pydantic_core>=2.41,<2.42
python-dotenv>=1.0,<1.1
python-multipart>=0.0.6,<0.0.7
PyYAML>=6.0,<6.1
regex>=2026.2,<2026.3
requests>=2.31,<2.32
requests-file>=3.0,<3.1
requests-toolbelt>=1.0,<1.1
sniffio>=1.3,<1.4
starlette>=0.35,<0.36
tenacity>=9.1,<9.2
tiktoken>=0.12,<0.13
tldextract>=5.3,<5.4
tqdm>=4.67,<4.68
typing-inspection>=0.4,<0.5
typing_extensions>=4.15,<4.16
urllib3>=2.6,<2.7
uuid_utils>=0.14,<0.15
uvicorn>=0.27,<0.28
uvloop>=0.22,<0.23
watchfiles>=1.1,<1.2
websockets>=16.0,<17.0
xxhash>=3.6,<3.7
yarl>=1.22,<1.23
zstandard>=0.25,<0.26
acontext>=0.1,<0.2
anthropic>=0.84,<0.85